CMake Dependency management
An attempt to consolidate the management of external dependencies to use
CMake rather than using git modules.

This is likely to be cleaner and more portable as you can rely on
pre-installed library packages when available and compile them only when
not available on the target development system

+++ ZEngine/include/ZEngine/Event/KeyReleasedEvent.h
@@ -20,7 +20,7 @@ namespace ZEngine::Event {
             return GetStaticCategory();
         }
         virtual std::string ToString() const override {
-            return fmt::format("KeyReleasedEvent : {0}", m_keycode);
+            return fmt::format("KeyReleasedEvent : {0}", static_cast<int>(m_keycode));
         }
     };
 } // namespace ZEngine::Event
